Displayed a more comprehensive error message if the base directory does not exist
Some FTP servers limit the access to a sub-directory of the filesystem while
other FTP servers expose the entire filesystem and could also start the FTP session
to the user home directory.
In this case the base directory is related to the root filesystem and not related
to the start of the TFP session.
This error message highlights the given base directory and the directory where the
FTP session started.